Support hierarchical dominant resource fairness
Hierarchical dominant resource fairness is configured with
a weighted tree, such that each node in the tree has a positive weight value.
The weights denote relative importance and do not have to sum to 1. The leaves
in the tree denote the jobs that are ultimately to be allocated resources,
whereas the internal nodes represent organizational or hierarchical groups.

In the case that an internal node has a child with a higher dominant share
while other children with another type of resource to be allocated will starve,
HDRF scales the child to the minimum dominant share and sums them up as the share of itself.

The hierarchy and weights of interest are annotated in the PodGroup
or Job annotations.

Signed-off-by: Peng Gao <peng.gao.dut@gmail.com>